requirements.txt
requests
python-gitlab